[FIX] fixed and issue when passing through expansion would lead to errors when trying to redirect from/in more than one file

+++ includes/expand.h
@@ -6,7 +6,7 @@
 /*   By: cviel <cviel@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/06/20 13:30:21 by cviel             #+#    #+#             */
-/*   Updated: 2025/08/01 00:34:01 by cviel            ###   ########.fr       */
+/*   Updated: 2025/08/04 16:55:33 by cviel            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -49,7 +49,7 @@ int		replace_key_value(t_exp *p_exp_part,
 			int *p_ind, int len_key, t_env *p_env);
 int		replace_err_code(t_exp *p_exp_part, int *p_ind, t_minishell mini);
 int		rebuild_elem(t_vector *dest, t_vector splitted);
-int		rebuild_filename(t_vector *dest, int ind, t_vector splitted);
+int		rebuild_filename(t_vector *dest, size_t ind, t_vector splitted);
 int		rebuild_exp(t_vector *dest, t_vector splitted, size_t ind);
 int		find_match(t_vector *p_names, t_vector pattern);
 int		matching(t_vector pattern, char *elem, int *p_match);
